move go-sdk integration tests to individual files
